Added the Offset rows
There are 3 types of offset rows. RETADDR for the return address space,
DISREG for the old display register space, FUNCVALUE for the return
function value.

These rows make it so the symbol table has the appropriate offset from
the Display register, after it is been setup by the caller and callee.