Callee Activation Records Complete
Function Return Value is removed from the symbol table. The symbol table
shows what the stack has from where the old display register is and up.
The function return value is part of the caller's record and not the
callee's activation record, thus it shouldn't be in the callee's symbol
table.

Analyzer now generates proper code to activated/deactivate on the
callee's side.