Added check into FormalParamSr
isCorrectParamCount returns if the pointer is equal to the size of the
formal parameter list. Which after being incremented in
Parser.actualParameter() it will be if there were equal actualParameters
and formalParameters in the code.

This check is performed in the lambda and the LParen sections of
optionalParameterList. If there were no parameters passed but should
have had parameters and if there were not enough parameters passed in.